Rename `Dish` and `DishComment` to occurrence
