+++ src/components/Occurrence/Occurrence.module.scss
