+++ src/components/Occurrence/comment/index.tsx
@@ -1,0 +1,3 @@
+import OccurrenceComment from './OccurrenceComment';
+
+export default OccurrenceComment;
