+++ src/components/Occurrence/index.tsx
@@ -1,0 +1,3 @@
+import Occurrence from './Occurrence';
+
+export default Occurrence;
