--- src/components/dish/comment/index.tsx
@@ -1,3 +1,0 @@
-import DishComment from './DishComment';
-
-export default DishComment;
