--- src/components/dish/index.tsx
@@ -1,3 +1,0 @@
-import Dish from './Dish';
-
-export default Dish;
